# backend/openapi.json
{
  "components": {
    "schemas": {
      "ActorRead": {
        "description": "Deliberately without `email`: an actor is identified by type and visible\nname, not by its credential. Keeps this schema agnostic of the actor's kind —\nwhen TASK-09 adds agents, the same schema serves them without changes.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "person",
              "agent"
            ],
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "display_name",
          "created_at"
        ],
        "title": "ActorRead",
        "type": "object"
      },
...
      "BoardCreate": {
        "properties": {
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "BoardCreate",
        "type": "object"
      },
      "BoardMemberCreate": {
        "properties": {
          "actor_id": {
            "format": "uuid",
            "title": "Actor Id",
            "type": "string"
          },
          "role": {
            "default": "member",
            "enum": [
              "owner",
              "member"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "actor_id"
        ],
        "title": "BoardMemberCreate",
        "type": "object"
      },
      "BoardMemberRead": {
        "properties": {
          "actor": {
            "$ref": "#/components/schemas/ActorRead"
          },
          "board_id": {
            "format": "uuid",
            "title": "Board Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "role": {
            "enum": [
              "owner",
              "member"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "id",
          "board_id",
          "actor",
          "role",
          "created_at"
        ],
        "title": "BoardMemberRead",
        "type": "object"
      },
      "BoardRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "role": {
            "enum": [
              "owner",
              "member"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "created_at",
          "role"
        ],
        "title": "BoardRead",
        "type": "object"
      },
      "BoardUpdate": {
        "properties": {
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "BoardUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_BoardMemberRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BoardMemberRead"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "size"
        ],
        "title": "Page[BoardMemberRead]",
        "type": "object"
      },
      "Page_BoardRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BoardRead"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "size"
        ],
        "title": "Page[BoardRead]",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "summary": "API de tableros kanban de kanbai.",
    "title": "kanbai",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/boards": {
      "get": {
        "operationId": "list_boards_api_v1_boards_get",
        "parameters": [
          {
            "description": "Página, empieza en 1",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "Página, empieza en 1",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Tamaño de página, máximo 100",
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Tamaño de página, máximo 100",
              "maximum": 100,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_BoardRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Lista los tableros de los que el actor autenticado es miembro",
        "tags": [
          "boards"
        ]
      },
      "post": {
        "operationId": "create_board_api_v1_boards_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BoardCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Crea un tablero nuevo",
        "tags": [
          "boards"
        ]
      }
    },
    "/api/v1/boards/{board_id}": {
      "delete": {
        "operationId": "delete_board_api_v1_boards__board_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "board_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Board Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "403": {
            "description": "Eres miembro, pero no owner."
          },
          "404": {
            "description": "No eres miembro de este tablero."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Borra un tablero (solo el owner)",
        "tags": [
          "boards"
        ]
      },
      "get": {
        "operationId": "get_board_api_v1_boards__board_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "board_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Board Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "No eres miembro de este tablero."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Obtiene un tablero por id",
        "tags": [
          "boards"
        ]
      },
      "patch": {
        "operationId": "rename_board_api_v1_boards__board_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "board_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Board Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BoardUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "description": "Eres miembro, pero no owner."
          },
          "404": {
            "description": "No eres miembro de este tablero."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Renombra un tablero (solo el owner)",
        "tags": [
          "boards"
        ]
      }
    },
...
    "/api/v1/boards/{board_id}/members": {
      "get": {
        "operationId": "list_members_api_v1_boards__board_id__members_get",
        "parameters": [
          {
            "in": "path",
            "name": "board_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Board Id",
              "type": "string"
            }
          },
          {
            "description": "Página, empieza en 1",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "Página, empieza en 1",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Tamaño de página, máximo 100",
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Tamaño de página, máximo 100",
              "maximum": 100,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_BoardMemberRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "No eres miembro de este tablero."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Lista los miembros de un tablero",
        "tags": [
          "boards"
        ]
      },
      "post": {
        "operationId": "add_member_api_v1_boards__board_id__members_post",
        "parameters": [
          {
            "in": "path",
            "name": "board_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Board Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BoardMemberCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardMemberRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "description": "Eres miembro, pero no owner."
          },
          "404": {
            "description": "No eres miembro de este tablero."
          },
          "409": {
            "description": "El actor ya es miembro del tablero, o es un agente cuya persona propietaria no tiene en este tablero un rol igual o superior."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Añade un miembro a un tablero (solo el owner; persona o agente por igual)",
        "tags": [
          "boards"
        ]
      }
    },
    "/api/v1/boards/{board_id}/members/{member_id}": {
      "delete": {
        "operationId": "remove_member_api_v1_boards__board_id__members__member_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "board_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Board Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "member_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Member Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "403": {
            "description": "Eres miembro, pero no owner."
          },
          "404": {
            "description": "No eres miembro de este tablero."
          },
          "409": {
            "description": "Es el único owner del tablero."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Quita a un miembro del tablero (el owner, o el propio miembro)",
        "tags": [
          "boards"
        ]
      }
    },
...
  }
}